simplify picker; fix compile errors
