+++ electricsql-quill/index.html
@@ -14,7 +14,7 @@
         height: calc(100% - 30px);
       }
     </style>
-    <title>ElectricSQL Recipe Editor</title>
+    <title>ElectricSQL-Quill</title>
   </head>
   <body>
     <div id="root"></div>
